requirements.txt
asgiref==3.5.2
click==8.1.3
colorama==0.4.5
Django==3.2.15
django-postgres-createdb==2020.12.3
django-rest-framework==0.1.0
djangorestframework==3.13.1
importlib-metadata==4.12.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
typed-ast==1.5.4
typing-extensions==4.3.0
zipp==3.8.1
